requirement.txt
requests
dateutil
pillow
python-thumbnails
sorl-thumbnail==12.3
